Update: Extract to util func
